fix: 좋아요 갯수조회 메서드 반환 타입 long -> int로 변경

+++ backend/src/main/java/com/ssafy/forest/repository/ReactionRepository.java
@@ -12,6 +12,6 @@ public interface ReactionRepository extends JpaRepository<Reaction, Long> {
 
     boolean existsByArticleIdAndMemberId(Long articleId, Long memberId);
 
-    Long countByArticleId(Long articleId);
+    int countByArticleId(Long articleId);
 
 }
